Move analysis types (ExpressionInfo, Reference, ReferenceFlags, ExpressionKind) from svelte_parser to svelte_analyze
These types are created and consumed entirely within svelte_analyze — the parser
never constructs them. Having them in svelte_parser pulled in oxc_semantic::SymbolId
as a parser dependency, violating the architectural boundary where symbol resolution
belongs to the analysis phase.
