+++ crates/svelte_analyze/src/lib.rs
@@ -22,9 +22,9 @@ pub(crate) mod walker;
 
 pub use data::{
     AnalysisData, AwaitBindingData, ClassDirectiveInfo, ComponentPropInfo, ComponentPropKind,
-    EventHandlerMode, LoweredTextPart, ConstTagData, ContentStrategy, DebugTagData, ElementFlags,
-    FragmentData, FragmentItem, FragmentKey, LoweredFragment, ParsedExprs, PropAnalysis, PropsAnalysis,
-    RenderTagCalleeMode, SnippetData,
+    EventHandlerMode, ExpressionInfo, ExpressionKind, LoweredTextPart, ConstTagData, ContentStrategy,
+    DebugTagData, ElementFlags, FragmentData, FragmentItem, FragmentKey, LoweredFragment, ParsedExprs,
+    PropAnalysis, PropsAnalysis, RenderTagCalleeMode, SnippetData,
 };
 pub use ident_gen::IdentGen;
 pub use scope::ComponentScoping;
