+++ crates/svelte_parser/src/types.rs
@@ -1,12 +1,11 @@
 //! Shared types and OXC utilities for the Svelte compiler.
 //!
-//! Leaf crate providing domain types (`ExpressionInfo`, `RuneKind`, `ScriptInfo`, etc.)
+//! Leaf crate providing domain types (`RuneKind`, `ScriptInfo`, etc.)
 //! and OXC parsing helpers used across parser, analyze, transform, and codegen.
 
 use oxc_ast::ast::Expression;
 
 use compact_str::CompactString;
-use oxc_semantic::SymbolId;
 use rustc_hash::{FxHashMap, FxHashSet};
 use svelte_ast::NodeId;
 use svelte_span::Span;
@@ -20,53 +19,6 @@ fn compact(s: &str) -> CompactString {
 // ---------------------------------------------------------------------------
 // Public types (owned, no lifetime)
 // ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone)]
-pub struct ExpressionInfo {
-    pub kind: ExpressionKind,
-    pub references: Vec<Reference>,
-    pub has_side_effects: bool,
-    pub has_call: bool,
-    /// Set when the expression contains `$effect.pending()` — forces the expression to be dynamic.
-    pub has_state_rune: bool,
-    /// Set when the expression contains a deep mutation on a `$`-prefixed identifier
-    /// (e.g., `$store.field = val` or `$store.count++`). Used to determine if component
-    /// needs `$.push/$.pop` for `$.store_mutate` support.
-    pub has_store_member_mutation: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct Reference {
-    pub name: CompactString,
-    pub span: Span,
-    pub flags: ReferenceFlags,
-    /// Resolved after `resolve_references` pass. `None` for globals/unresolved.
-    pub symbol_id: Option<SymbolId>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ReferenceFlags {
-    Read,
-    Write,
-    ReadWrite,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ExpressionKind {
-    Identifier(CompactString),
-    Literal,
-    CallExpression { callee: CompactString },
-    MemberExpression,
-    ArrowFunction,
-    Assignment,
-    Other,
-}
-
-impl ExpressionKind {
-    pub fn is_simple(&self) -> bool {
-        matches!(self, Self::Identifier(_) | Self::MemberExpression)
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct PropInfo {
